tutorials/Learning Series 10/requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.8
lifelines>=0.27
scikit-learn>=1.3
pyyaml>=6.0
